Refactorización de importaciones: traslado de FIFTEEN_MINUTES_MS a utils/checkInUtils para mejorar la organización.
